+++ Tests/VCoreTests/Tests/Extensions/Foundation/StringDiacriticInsensitiveStringTests.swift
@@ -1,0 +1,16 @@
+//
+//  StringDiacriticInsensitiveStringTests.swift
+//  VCoreTests
+//
+//  Created by Vakhtang Kontridze on 13.07.22.
+//
+
+import XCTest
+@testable import VCore
+
+// MARK: - String Diacritic Insensitive String Tests
+final class StringDiacriticInsensitiveStringTests: XCTestCase {
+    func test() {
+        XCTAssertEqual("À".diacriticInsensitiveString(), "A")
+    }
+}