+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Views and ViewControllers
 Extensions
 
 - `Double.rounded(fractions:)` and `Double.round(fractions:)` methods are added
+- `String.diacriticInsensitiveString(locale:)` method is added
 
 ### [3.14.0(45)](https://github.com/VakhoKontridze/VCore/releases/tag/3.14.0) — *2022 07 12*